refactor: extract formatters to separate module to resolve circular imports (#20)
